requirements.txt
//beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
requests==2.31.0
tqdm==4.66.1
